Rename RobotMeasurements to PlayerMeasurements
Add comments
